requirements.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
yfinance==0.2.20
ta==0.10.2
plotly==5.17.0
matplotlib==3.7.2
quantstats==0.0.59
fpdf2==2.7.5
python-dateutil==2.8.2
